Update row and column constraints
Instead of having the "if" behaviour in the Python scripting, we should
have it in the propositional logic itself. This will allow us to get
more insight into the problem from the SAT solver.

Additionally, syntax issues were corrected (`x[i][j+0]` -> `x[i][0]`)
which caused issues when attempting to run the file.